Big Change
Bisschen übersichtlicher. Neue Klasse Entity und Tower. Troops sind jetzt nicht mehr alles, sondern nur eine Form von Entitys
