+++ Tower.java
@@ -1,0 +1,9 @@
+public class Tower extends  Entity{
+
+    Tower(Card card, double x, double y, Spieler affiliation) {
+        super(card, x, y, affiliation);
+    }
+    void KickTheBucket(){
+
+    }
+}
